requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
discord.py==1.2.4
idna==2.8
isort==4.3.21
joblib==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.1
mccabe==0.6.1
multidict==4.5.2
numpy==1.17.3
pandas==0.25.3
pylint==2.4.3
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
sklearn==0.0
typed-ast==1.4.0
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
